requirements.txt
aiogram==3.7.0
APScheduler==3.10.4
SQLAlchemy>=1.4,<3.0
aiosqlite==0.20.0
aiohttp==3.9.5
google-auth==2.29.0
google-api-python-client==2.127.0
google-auth-httplib2==0.2.0
Pillow==10.3.0
python-dotenv==1.0.1
openai==1.30.0
pytz>=2024.1
